% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmix-package.R
\docType{package}
\name{bbmix-package}
\alias{bbmix-package}
\alias{bbmix}
\title{The 'bbmix' package.}
\description{
Bayesian Beta-Binomial mixture model for RNA-seq genotyping
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. https://mc-stan.org
}
