% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmix.R
\name{call_help}
\alias{call_help}
\title{call gt helper, calculate mean dbetabinom from all posterior samples}
\usage{
call_help(n, m, mu, lambda)
}
\arguments{
\item{n}{counts alt allele}

\item{m}{total counts}

\item{mu}{vector with posterior draws for mu param}

\item{lambda}{vector with posterior draws for lambda param}
}
\value{
mean of dbetabinom
}
\description{
call gt helper, calculate mean dbetabinom from all posterior samples
}
