% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmix.R
\name{ex_alt_hom}
\alias{ex_alt_hom}
\title{Exclude fSNPs with no alternative allele in any sample. Also exclude fSNPs if all samples are hom.}
\usage{
ex_alt_hom(gt_f, out)
}
\arguments{
\item{gt_f}{character vector with file names with genotype calls per sample}

\item{out}{file name to save output}
}
\value{
save file
}
\description{
Exclude fSNPs with no alternative allele in any sample. Also exclude fSNPs if all samples are hom.
}
\examples{

gt_f <- system.file("extdata/output", "gt.NA12878.chr22.txt",
package = "bbmix",
mustWork = TRUE)
out <- tempfile()

## Running function
ex_alt_hom(gt_f, out)

unlink(out)

}
