% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmix.R
\name{gt_help}
\alias{gt_help}
\title{call gt helper, get posterior mean, expected gt and sd expected gt across all samples}
\usage{
gt_help(stan_samples, pop, data)
}
\arguments{
\item{stan_samples}{matrix with samples extracted from stan fit object, params mu and lambda}

\item{pop}{population to select AF for GT prior, defaults to EUR}

\item{data}{data table 1 row with counts and EAF to apply model}
}
\value{
gt_help()
}
\description{
call gt helper, get posterior mean, expected gt and sd expected gt across all samples
}
